# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Онлайн-кинотеатр API"
  description: |
    Полноценный API для стримингового сервиса с поддержкой фильмов, сериалов, рейтингов и пользовательских закладок.
    
    ## Основные возможности:
    - 📺 Просмотр фильмов и сериалов
    - ⭐ Система рейтингов и отзывов
    - 📋 Личные закладки пользователей
    - 👥 Управление пользователями и подписками
    - 📊 Аналитика и статистика
    - 🔧 Управление справочниками (админ)
    
    ## Аутентификация:
    Все защищенные эндпоинты требуют JWT токен, полученный через `/auth/login`.
    
    ## Технологии:
    - FastAPI 0.104+
    - SQLModel (SQLAlchemy + Pydantic)
    - JWT аутентификация
    - PostgreSQL / SQLite

  version: "2.0.0"
  contact:
    name: "API Support"
    email: "support@cinema.example.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://localhost:8000"
    description: "Локальный сервер разработки"
  - url: "https://api.cinema.example.com"
    description: "Продакшн сервер"
...
paths:
...
  /content/:
    get:
      summary: "Получить список контента"
      description: |
        Получение списка фильмов и сериалов с пагинацией и фильтрацией.
        
        **Фильтры:**
        - По типу контента (фильм/сериал)
        - По жанру
        - По году выпуска
        - По возрастному рейтингу
        - По минимальному рейтингу IMDB
      tags: ["Контент"]
      parameters:
        - name: content_type_id
          in: query
          required: false
          description: "ID типа контента (из справочника)"
          schema:
            type: integer
        - name: genre_id
          in: query
          required: false
          description: "ID жанра"
          schema:
            type: integer
        - name: min_year
          in: query
          required: false
          description: "Минимальный год выпуска"
          schema:
            type: integer
            minimum: 1900
        - name: max_year
          in: query
          required: false
          description: "Максимальный год выпуска"
          schema:
            type: integer
            maximum: 2100
        - name: min_rating
          in: query
          required: false
          description: "Минимальный рейтинг IMDB"
          schema:
            type: number
            minimum: 0
            maximum: 10
        - name: limit
          in: query
          required: false
          description: "Количество элементов на странице"
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: "Смещение для пагинации"
          schema:
            type: integer
            minimum: 0
            default: 0
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Список контента"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentResponse"
        "401":
          description: "Неавторизованный доступ"
  /content/{content_id}:
    get:
      summary: "Получить детальную информацию о контенте"
      description: "Получение полной информации о фильме или сериале"
      tags: ["Контент"]
      parameters:
        - name: content_id
          in: path
          required: true
          description: "ID контента"
          schema:
            type: integer
            minimum: 1
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Детальная информация о контенте"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetailResponse"
        "401":
          description: "Неавторизованный доступ"
        "403":
          description: "Возрастные ограничения или ограничения подписки"
        "404":
          description: "Контент не найден"
  /content/{content_id}/episodes:
    get:
      summary: "Получить эпизоды сериала"
      description: "Получение списка эпизодов для сериала"
      tags: ["Контент"]
      parameters:
        - name: content_id
          in: path
          required: true
          description: "ID контента (сериала)"
          schema:
            type: integer
            minimum: 1
        - name: season_number
          in: query
          required: false
          description: "Номер сезона (если нужен конкретный сезон)"
          schema:
            type: integer
            minimum: 1
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Список эпизодов, сгруппированных по сезонам"
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/SeasonWithEpisodes"
        "401":
          description: "Неавторизованный доступ"
        "404":
          description: "Сериал не найден"
...
  /content/featured/today:
    get:
      summary: "Сегодняшние рекомендации"
      description: "Получить рекомендованный контент на сегодня"
      tags: ["Контент"]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Список рекомендованного контента"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentResponse"
        "401":
          description: "Неавторизованный доступ"

  # ============ Рейтинги ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен аутентификации"
  schemas:
...
    ContentBase:
      type: object
      required: [title, content_type_id, age_rating_id, release_year]
      properties:
        title:
          type: string
          maxLength: 255
          example: "Интерстеллар"
        description:
          type: string
          nullable: true
          example: "Фантастический эпос о путешествии сквозь червоточину"
        content_type_id:
          type: integer
          example: 1
        age_rating_id:
          type: integer
          example: 3
        release_year:
          type: integer
          minimum: 1900
          maximum: 2100
          example: 2014
        duration_minutes:
          type: integer
          minimum: 0
          example: 169
        director:
          type: string
          nullable: true
          example: "Кристофер Нолан"
        imdb_rating:
          type: number
          minimum: 0
          maximum: 10
          nullable: true
          example: 8.6
...
    ContentResponse:
      allOf:
        - $ref: "#/components/schemas/ContentBase"
        - type: object
          required: [id, is_featured, available_from]
          properties:
            id:
              type: integer
              example: 1
            poster_url:
              type: string
              format: uri
              nullable: true
            trailer_url:
              type: string
              format: uri
              nullable: true
            is_featured:
              type: boolean
              example: true
            available_from:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
            content_type:
              $ref: "#/components/schemas/ContentTypeResponse"
            age_rating:
              $ref: "#/components/schemas/AgeRatingResponse"
    ContentDetailResponse:
      allOf:
        - $ref: "#/components/schemas/ContentResponse"
        - type: object
          properties:
            genres:
              type: array
              items:
                $ref: "#/components/schemas/GenreResponse"
            actors:
              type: array
              items:
                $ref: "#/components/schemas/ActorResponse"
            seasons:
              type: array
              items:
                $ref: "#/components/schemas/SeasonResponse"
            ratings_summary:
              $ref: "#/components/schemas/RatingsSummary"

    # ============ Жанры и актеры ============
    GenreResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Научная фантастика"
        description:
          type: string
          nullable: true
          example: "Фантастика, основанная на научных концепциях"
    ActorResponse:
      type: object
      required: [id, full_name]
      properties:
        id:
          type: integer
          example: 1
        full_name:
          type: string
          example: "Мэттью Макконахи"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1969-11-04"
        biography:
          type: string
          nullable: true
        photo_url:
          type: string
          format: uri
          nullable: true

    # ============ Сезоны и эпизоды ============
    SeasonResponse:
      type: object
      required: [id, season_number]
      properties:
        id:
          type: integer
          example: 1
        season_number:
          type: integer
          minimum: 1
          example: 1
        title:
          type: string
          nullable: true
          example: "Первый сезон"
        release_year:
          type: integer
          nullable: true
          example: 2023
    SeasonWithEpisodes:
      type: object
      properties:
        season_info:
          $ref: "#/components/schemas/SeasonResponse"
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/EpisodeResponse"
    EpisodeResponse:
      type: object
      required: [id, episode_number, title, duration_minutes, video_url]
      properties:
        id:
          type: integer
          example: 1
        episode_number:
          type: integer
          minimum: 1
          example: 5
        title:
          type: string
          example: "Поворотный момент"
        description:
          type: string
          nullable: true
        duration_minutes:
          type: integer
          minimum: 0
          example: 47
        video_url:
          type: string
          format: uri
          example: "https://cdn.cinema.com/episodes/s01e05.mp4"
        thumbnail_url:
          type: string
          format: uri
          nullable: true
        release_date:
          type: string
          format: date
          nullable: true
          example: "2023-11-15"
        video_quality:
          $ref: "#/components/schemas/VideoQualityResponse"

    # ============ Рейтинги ============
...
    RatingsSummary:
      type: object
      properties:
        total_ratings:
          type: integer
          example: 125
        average_rating:
          type: number
          example: 4.5
        min_rating:
          type: number
          example: 1.0
        max_rating:
          type: number
          example: 5.0
...
    ContentTypeResponse:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          example: 1
        code:
          type: string
          example: "movie"
        name:
          type: string
          example: "Фильм"
        description:
          type: string
          nullable: true
        icon_url:
          type: string
          format: uri
          nullable: true
        sort_order:
          type: integer
          example: 1
    AgeRatingResponse:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          example: 3
        code:
          type: string
          example: "PG-13"
        name:
          type: string
          example: "С 13 лет"
        min_age:
          type: integer
          nullable: true
          example: 13
        description:
          type: string
          nullable: true
        icon_url:
          type: string
          format: uri
          nullable: true
...
    VideoQualityResponse:
      type: object
      required: [id, code, name, width, height]
      properties:
        id:
          type: integer
          example: 3
        code:
          type: string
          example: "fullhd"
        name:
          type: string
          example: "Full HD"
        width:
          type: integer
          example: 1920
        height:
          type: integer
          example: 1080
        bitrate:
          type: integer
          nullable: true
          example: 5000